+++ test.mjs
@@ -966,6 +966,15 @@ aa();
             scope: [
                 "(function aa(bb = aa) {\n    aa(bb);\n}());",
                 "function aa(bb = aa) {\n    aa(bb);\n}\naa();",
+                //!! (`
+//!! if (String) {
+    //!! let aa = 0;
+    //!! aa();
+//!! } else {
+    //!! let aa = 0;
+    //!! aa();
+//!! }
+                //!! `),
                 (`
 if (String) {
     var aa = 0; //jslint-ignore-line