+++ CHANGELOG.md
@@ -1,6 +1,19 @@
 # Changelog
 
 # Todo
+- jslint - Rename token-property for variables '.init' to '.assigned', to improve readability.
+- jslint - Rename internal variable 'blockage' to 'block_scope', 'functionage' to 'function_scope', to imporove readability.
+- jslint - Change warning 'uninitialized_a' to 'unassigned_variable_a'.
+- jslint - Change warning 'out_of_scope_a' to 'temporal_dead_zone_a'.
+- jslint - Add warning and tdz for function-declaration inside block-scope.
+- jslint - Add implicit block-scope for:
+    - if-else
+    - for-loop
+        - for-semicolon
+        - for-in
+        - for-of
+    - while-loop
+    - do-while
 - jslint-ecma - Expand ES2015-feature-support for es-module-export-statement.
 - jslint - Relax warning expected_line_break_a_b for ternary-operator inside template-literal.
 - jslint - Add ability to auto-fix whitespace.
@@ -9,8 +22,15 @@
 - jslint - Try to improve parser to be able to parse jquery.js without stopping.
 
 # v2026.7.1-beta
-- jslint - Move catch-variable from its special-scope into lexical-scope.
-- jslint - Add lexical-scope to internal-function jslint_phase3_parse().
+- jslint - Change scope from function-scope to block-scope:
+    - const-declaration
+    - let-declaration
+    - function-declaration
+- jslint - Restrict scope from function-scope to its own function-body:
+    - named-function-expression
+- jslint - Change scope from special-catch-scope to block-scope:
+    - catch-variable
+- jslint - Add block-scope to internal-function jslint_phase3_parse().
 - jslint-ecma - Add ES2015-feature for..of.
 - jslint - Expand built-in-globals for browser, ecma, and node - auto-generated from online-sources.
 - jslint-ci - Add automated ci for shellcheck to lint shell-scripts.
@@ -21,8 +41,8 @@
 - jslint-regression - Fix long-running regression where 'let x = x;' doesn't warn about temporal-dead-zone.
 - jslint-warning - Tighten warning of unused variables to be always on, regardless of module / nodejs mode.
 - jslint - Wrap all property-updates 'name.init = true/false' with calls to:
-    name_lookup() - 'aa=0'
-    name_push()   - 'let aa=0'
+    name_declare()      - 'let aa=0'
+    post_a_assignment() - 'aa=0'
 
 # v2026.6.30
 - jslint-ecma - Update README.md, documenting supported ES2015+ features.